# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "My API",
    "description": "This is a sample API documentation",
    "version": "1.0.0",
    "contact": {
      "name": "SilkyWay API Support",
      "url": "http://www.silkyway.vercel.app",
      "email": "silkyway.enterprise@gmail.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local server"
    },
    {
      "url": "https://silkyway.vercel.app",
      "description": "Production server"
    }
  ],
  "tags": [
    {
      "name": "Categories",
      "description": "Operations about categories"
    },
    {
      "name": "Products",
      "description": "Operations about products"
    },
    {
      "name": "Users",
      "description": "Operations about users"
    },
    {
      "name": "Authentication",
      "description": "Operations about authentication"
    }
  ],
  "paths": {
    "/api/products": {
      "get": {
        "tags": ["Products"],
        "summary": "Get list of products",
        "responses": {
          "200": {
            "description": "A list of products",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": ["Products"],
        "summary": "Create a new product",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Name of the product",
                    "example": "Sample Product"
                  },
                  "description": {
                    "type": "string",
                    "description": "Description of the product",
                    "example": "This is a sample product."
                  },
                  "price": {
                    "type": "number",
                    "description": "Price of the product",
                    "example": 19.99
                  },
                  "stock": {
                    "type": "number",
                    "description": "Stock of the product",
                    "example": 100
                  },
                  "images": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "description": "URLs of product images",
                      "example": "http://example.com/image1.jpg"
                    },
                    "description": "Images of the product"
                  },
                  "categoryID": {
                    "type": "string",
                    "description": "ID of the product category",
                    "example": "60d6e88f53112345678d960c"
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "description": "Tags associated with the product",
                      "example": "new, sale, popular"
                    },
                    "description": "Tags of the product"
                  }
                },
                "required": ["name", "price", "images"]
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Product created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "409": {
            "description": "Conflict - Product with the same name already exists"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/products/getProductsByFilter": {
      "get": {
        "tags": ["Products"],
        "summary": "Get list of products by given parameters",
        "parameters": [
          {
            "name": "priceRange",
            "in": "query",
            "description": "Price range filter (min-max)",
            "required": false,
            "schema": {
              "type": "string",
              "example": "10-50"
            }
          },
          {
            "name": "tag",
            "in": "query",
            "description": "A relevant Tag",
            "required": false,
            "schema": {
              "type": "string",
              "example": "Jewellery"
            }
          },
          {
            "name": "categoryID",
            "in": "query",
            "description": "Category filter",
            "required": false,
            "schema": {
              "type": "string",
              "example": "1"
            }
          },
          {
            "name": "name",
            "in": "query",
            "description": "Name filter",
            "required": false,
            "schema": {
              "type": "string",
              "example": "Sample"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of products",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Product not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/products/{id}": {
  "get": {
    "tags": ["Products"],
    "summary": "Get a product by ID",
    "parameters": [
      {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "example": "60d6e88f53112345678d960c"
      }
    ],
    "responses": {
      "200": {
        "description": "Product details",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Product"
            }
          }
        }
      },
      "404": {
        "description": "Product not found"
      },
      "500": {
        "description": "Internal server error"
      }
    }
  },
  "put": {
    "tags": ["Products"],
    "summary": "Update a product",
    "description": "Update the name, price, and stock of a product",
    "parameters": [
      {
        "name": "id",
        "in": "path",
        "description": "ID of the product to update",
        "required": true,
        "schema": {
          "type": "string"
        }
      }
    ],
    "requestBody": {
      "description": "New name, price, and stock of the product",
      "required": true,
      "content": {
        "application/json": {
          "schema": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "description": "New name of the product"
              },
              "price": {
                "type": "number",
                "description": "New price of the product"
              },
              "stock": {
                "type": "number",
                "description": "New stock of the product"
              }
            },
            "required": ["name", "price", "stock"]
          }
        }
      }
    },
    "responses": {
      "200": {
        "description": "Product updated successfully",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Product"
            }
          }
        }
      },
      "400": {
        "description": "All fields are required"
      },
      "404": {
        "description": "Product not found"
      },
      "500": {
        "description": "Internal server error"
      }
    }
  },
  "delete": {
    "tags": ["Products"],
    "summary": "Delete a product by ID",
    "parameters": [
      {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "example": "60d6e88f53112345678d960c"
      }
    ],
    "responses": {
      "204": {
        "description": "Product deleted"
      },
      "404": {
        "description": "Product not found"
      },
      "500": {
        "description": "Internal server error"
      }
    }
  }
},
    "/api/categories": {
      "get": {
        "tags": ["Categories"],
        "summary": "Get list of categories",
        "responses": {
          "200": {
            "description": "A list of categories",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": ["Categories"],
        "summary": "Create a new category",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Category"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Category created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/users": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Get list of users",
        "responses": {
          "200": {
            "description": "A list of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": ["Users"],
        "summary": "Create a new user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "api/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Login an existing user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "user@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "password123"
                  }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User successfully logged in",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "$ref": "#/components/schemas/User"
                    },
                    "token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid email/password"
          },
          "401": {
            "description": "Unauthorized - Incorrect email/password"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "api/auth/register": {
      "post": {
        "tags": ["Users"],
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegister"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRegister"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Email address already in use or invalid data format"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "tags": ["Users"],
        "summary": "Get a user by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "User details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "example": 1
                    },
                    "name": {
                      "type": "string",
                      "example": "John Doe"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "put": {
        "tags": ["Users"],
        "summary": "Update a user by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "example": 1
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "John Smith"
                  }
                },
                "required": ["name"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "example": 1
                    },
                    "name": {
                      "type": "string",
                      "example": "John Smith"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "delete": {
        "tags": ["Users"],
        "summary": "Delete a user by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "example": 1
          }
        ],
        "responses": {
          "204": {
            "description": "User deleted"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Product": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "required": true,
            "trim": true,
            "example": "Sample Product"
          },
          "description": {
            "type": "string",
            "required": true,
            "example": "This is a sample product description"
          },
          "price": {
            "type": "number",
            "example": 19.99
          },
          "categoryID": {
            "type": "number",
            "example": 1
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "stock": {
            "type": "number",
            "example": 100
          },
          "images": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "url": {
                  "type": "string",
                  "example": "http://example.com/image1.jpg"
                },
                "alt": {
                  "type": "string",
                  "example": "Image description"
                }
              }
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2023-06-01T00:00:00Z"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "example": "2023-06-01T00:00:00Z"
          }
        }
      },
      "Category": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "required": true,
            "trim": true,
            "example": "Electronics"
          },
          "description": {
            "type": "string",
            "example": "Category description"
          },
          "categoryID": {
            "type": "number",
            "required": true,
            "example": 1
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2023-06-01T00:00:00Z"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "example": "2023-06-01T00:00:00Z"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "required": true,
            "unique": true,
            "example": "user@example.com"
          },
          "name": {
            "type": "string",
            "required": true,
            "trim": true,
            "example": "John Doe"
          },
          "password": {
            "type": "string",
            "required": true,
            "minlength": 6,
            "example": "password123"
          }
        }
      },
      "UserRegister": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "required": true,
            "unique": true,
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "required": true,
            "minlength": 6,
            "example": "password123"
          },
          "firstname": {
            "type": "string",
            "required": true,
            "trim": true,
            "example": "John"
          },
          "lastname": {
            "type": "string",
            "required": true,
            "trim": true,
            "example": "Doe"
          },
          "address": {
            "type": "string",
            "required": true,
            "trim": true,
            "example": "example@hotmail.com"
          }
        }
      }
    }
  }
}